Finished all stages, excluding the editing func

+++ Lab3/main.h
@@ -8,7 +8,7 @@ typedef struct stage_t {
 	uint8_t timeLen;
 	uint8_t totalLen;     // concat options with time
 	int8_t  highlight;    // index (out of total)l; -1 for none
-	uint8_t color[5];     // 1 for highlighted
+	int color[5];     
 	uint8_t selected;      // 0 for not selected. 1 for selected
 } stage;
